Gismo: change "Assignments" view to pull grades from gradebook rather than assignment submissions table
As assignments can be marked through the gradebook rather than the assignment submissions screen, grades should be pulled from the gradebook rather than the assignment submissions table.  The latter also appears to truncate grades, so fractional marks are ignored and rounded down - switching to the gradebook also deals with this.  Gradebook "final" grades are rounded to 1dp for display in Gismo.